Revert "run payroll now works"
This reverts commit 4ac598411da46910abcd1ed0bba792a808bc6be4.

+++ javaSide/JavaPayroll/src/controll/Database.java
@@ -153,14 +153,14 @@ public class Database {
 		}		
 	}
 
-	public void addPayHeld(int id, float amount, String date) {
+	public void addPayHeld(int id, int amount, String date) {
 		if (isConnected) {
 			try {
 				String sql = "INSERT INTO `held_pay`(`empId`, `amount`, `date`) VALUES " + 
 						"(?,?);";
 				java.sql.PreparedStatement statement = connection.prepareStatement(sql);
 				statement.setInt(1, id);
-				statement.setFloat(2, amount);
+				statement.setInt(2, amount);
 				statement.setString(3, date);
 				statement.executeUpdate();
 			} catch (SQLException e) {
@@ -171,7 +171,7 @@ public class Database {
 	}
 
 	public void addMailPay(String firstName, String lastName, String address,
-			float amount, int id, String date) {
+			int amount, int id, String date) {
 		if (isConnected) {
 			try {
 				String sql = "INSERT INTO `mail_pay`(`empId`, `firstName`, `lastName`, `address`, `amount`, `date`) VALUES " + 
@@ -181,7 +181,7 @@ public class Database {
 				statement.setString(2, firstName);
 				statement.setString(3, lastName);
 				statement.setString(4, address);
-				statement.setFloat(5, amount);
+				statement.setInt(5, amount);
 				statement.setString(6, date);
 				statement.executeUpdate();
 			} catch (SQLException e) {
@@ -191,7 +191,7 @@ public class Database {
 		}			
 	}
 
-	public void addDepositePay(String firstName, String lastName, float amount, int empId, String date) {
+	public void addDepositePay(String firstName, String lastName, int amount, int empId, String date) {
 		if (isConnected) {
 			try {
 				String sql = "INSERT INTO `direct_pay`(`empId`, `firstName`, `lastName`, `account`, `amount`, `date`) VALUES " + 
@@ -201,7 +201,7 @@ public class Database {
 				statement.setString(2, firstName);
 				statement.setString(3, lastName);
 				statement.setString(4, "00000000");
-				statement.setFloat(5, amount);
+				statement.setInt(5, amount);
 				statement.setString(6, date);
 				statement.executeUpdate();
 			} catch (SQLException e) {
